LPS-51926 Move DL tests to its bundle and make them OSGi friendly
