+++ modules/apps/document-library/document-library-test/test/integration/com/liferay/document/library/lar/test/DLFileEntryTypeStagedModelDataHandlerTest.java
@@ -12,29 +12,25 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.lar;
+package com.liferay.document.library.lar.test;
 
-import com.liferay.portal.kernel.repository.model.Folder;
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.dynamic.data.mapping.model.DDMStructure;
+import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.lar.test.BaseStagedModelDataHandlerTestCase;
 import com.liferay.portal.model.Group;
-import com.liferay.portal.model.Repository;
-import com.liferay.portal.model.RepositoryEntry;
 import com.liferay.portal.model.StagedModel;
-import com.liferay.portal.repository.liferayrepository.LiferayRepository;
-import com.liferay.portal.service.RepositoryEntryLocalServiceUtil;
-import com.liferay.portal.service.RepositoryLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.MainServletTestRule;
-import com.liferay.portal.util.PortalUtil;
-import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
-import com.liferay.portlet.documentlibrary.service.DLAppServiceUtil;
+import com.liferay.portlet.documentlibrary.model.DLFileEntryMetadata;
+import com.liferay.portlet.documentlibrary.model.DLFileEntryType;
+import com.liferay.portlet.documentlibrary.service.DLFileEntryTypeLocalServiceUtil;
+import com.liferay.portlet.dynamicdatamapping.DDMStructureManagerUtil;
 
 import java.util.HashMap;
 import java.util.List;
@@ -43,19 +39,20 @@ import java.util.Map;
 import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
  * @author Mate Thurzo
  */
-public class RepositoryStagedModelDataHandlerTest
+@RunWith(Arquillian.class)
+public class DLFileEntryTypeStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
-			TransactionalTestRule.INSTANCE);
+			new LiferayIntegrationTestRule(), TransactionalTestRule.INSTANCE);
 
 	@Override
 	protected Map<String, List<StagedModel>> addDependentStagedModelsMap(
@@ -65,32 +62,12 @@ public class RepositoryStagedModelDataHandlerTest
 		Map<String, List<StagedModel>> dependentStagedModelsMap =
 			new HashMap<>();
 
-		long classNameId = PortalUtil.getClassNameId(
-			LiferayRepository.class.getName());
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				group.getGroupId(), TestPropsValues.getUserId());
-
-		Folder mountFolder = DLAppServiceUtil.addFolder(
-			group.getGroupId(), DLFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			serviceContext);
-
-		_repository = RepositoryLocalServiceUtil.addRepository(
-			TestPropsValues.getUserId(), group.getGroupId(), classNameId,
-			mountFolder.getFolderId(), RandomTestUtil.randomString(),
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			new UnicodeProperties(), false, serviceContext);
-
-		RepositoryEntry repositoryEntry =
-			RepositoryEntryLocalServiceUtil.addRepositoryEntry(
-				TestPropsValues.getUserId(), group.getGroupId(),
-				_repository.getRepositoryId(), RandomTestUtil.randomString(),
-				serviceContext);
+		DDMStructure ddmStructure = DDMStructureTestUtil.addStructure(
+			group.getGroupId(), DLFileEntryMetadata.class.getName());
 
 		addDependentStagedModel(
-			dependentStagedModelsMap, RepositoryEntry.class, repositoryEntry);
+			dependentStagedModelsMap,
+			DDMStructureManagerUtil.getDDMStructureModelClass(), ddmStructure);
 
 		return dependentStagedModelsMap;
 	}
@@ -101,14 +78,29 @@ public class RepositoryStagedModelDataHandlerTest
 			Map<String, List<StagedModel>> dependentStagedModelsMap)
 		throws Exception {
 
-		return _repository;
+		Class<?> ddmStructureClass =
+			DDMStructureManagerUtil.getDDMStructureModelClass();
+
+		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
+			ddmStructureClass.getSimpleName());
+
+		DDMStructure ddmStructure = (DDMStructure)dependentStagedModels.get(0);
+
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				group.getGroupId(), TestPropsValues.getUserId());
+
+		return DLFileEntryTypeLocalServiceUtil.addFileEntryType(
+			TestPropsValues.getUserId(), group.getGroupId(),
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			new long[] {ddmStructure.getStructureId()}, serviceContext);
 	}
 
 	@Override
 	protected StagedModel getStagedModel(String uuid, Group group) {
 		try {
-			return RepositoryLocalServiceUtil.getRepositoryByUuidAndGroupId(
-				uuid, group.getGroupId());
+			return DLFileEntryTypeLocalServiceUtil.
+				getDLFileEntryTypeByUuidAndGroupId(uuid, group.getGroupId());
 		}
 		catch (Exception e) {
 			return null;
@@ -117,7 +109,7 @@ public class RepositoryStagedModelDataHandlerTest
 
 	@Override
 	protected Class<? extends StagedModel> getStagedModelClass() {
-		return Repository.class;
+		return DLFileEntryType.class;
 	}
 
 	@Override
@@ -126,18 +118,18 @@ public class RepositoryStagedModelDataHandlerTest
 			Group group)
 		throws Exception {
 
+		Class<?> ddmStructureClass =
+			DDMStructureManagerUtil.getDDMStructureModelClass();
+
 		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
-			RepositoryEntry.class.getSimpleName());
+			ddmStructureClass.getSimpleName());
 
 		Assert.assertEquals(1, dependentStagedModels.size());
 
-		RepositoryEntry repositoryEntry =
-			(RepositoryEntry)dependentStagedModels.get(0);
+		DDMStructure ddmStructure = (DDMStructure)dependentStagedModels.get(0);
 
-		RepositoryEntryLocalServiceUtil.getRepositoryEntryByUuidAndGroupId(
-			repositoryEntry.getUuid(), group.getGroupId());
+		DDMStructureManagerUtil.getStructureByUuidAndGroupId(
+			ddmStructure.getUuid(), group.getGroupId());
 	}
-
-	private Repository _repository;
 
 }
