+++ modules/apps/document-library/document-library-test/test/integration/com/liferay/document/library/lar/test/FolderStagedModelDataHandlerTest.java
@@ -12,45 +12,40 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.lar;
+package com.liferay.document.library.lar.test;
 
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.repository.model.FileVersion;
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.dynamic.data.mapping.model.DDMStructure;
+import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
 import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.DateUtil;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.lar.test.BaseStagedModelDataHandlerTestCase;
 import com.liferay.portal.model.Company;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.StagedModel;
+import com.liferay.portal.repository.liferayrepository.model.LiferayFolder;
 import com.liferay.portal.service.CompanyLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.ServiceContextThreadLocal;
-import com.liferay.portal.test.randomizerbumpers.TikaSafeRandomizerBumper;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.MainServletTestRule;
-import com.liferay.portlet.documentlibrary.model.DLFileEntry;
 import com.liferay.portlet.documentlibrary.model.DLFileEntryMetadata;
 import com.liferay.portlet.documentlibrary.model.DLFileEntryType;
 import com.liferay.portlet.documentlibrary.model.DLFolder;
 import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
-import com.liferay.portlet.documentlibrary.service.DLAppLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.service.DLAppServiceUtil;
 import com.liferay.portlet.documentlibrary.service.DLFileEntryTypeLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.service.DLFolderLocalServiceUtil;
-import com.liferay.portlet.documentlibrary.util.test.DLAppTestUtil;
-import com.liferay.portlet.dynamicdatamapping.DDMStructure;
 import com.liferay.portlet.dynamicdatamapping.DDMStructureManagerUtil;
-import com.liferay.portlet.dynamicdatamapping.util.test.DDMStructureTestUtil;
+import com.liferay.portlet.exportimport.lar.StagedModelDataHandler;
+import com.liferay.portlet.exportimport.lar.StagedModelDataHandlerRegistryUtil;
+import com.liferay.portlet.exportimport.lar.StagedModelDataHandlerUtil;
 
+import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -58,69 +53,44 @@ import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Mate Thurzo
  */
-public class FileEntryStagedModelDataHandlerTest
+@RunWith(Arquillian.class)
+public class FolderStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
-			TransactionalTestRule.INSTANCE);
+			new LiferayIntegrationTestRule(), TransactionalTestRule.INSTANCE);
 
 	@Test
 	public void testCompanyScopeDependencies() throws Exception {
+		initExport();
+
 		Map<String, List<StagedModel>> dependentStagedModelsMap =
 			addCompanyDependencies();
 
 		StagedModel stagedModel = addStagedModel(
 			stagingGroup, dependentStagedModelsMap);
 
-		exportImportStagedModel(stagedModel);
+		StagedModelDataHandlerUtil.exportStagedModel(
+			portletDataContext, stagedModel);
+
+		initImport();
+
+		StagedModel exportedStagedModel = readExportedStagedModel(stagedModel);
+
+		Assert.assertNotNull(exportedStagedModel);
+
+		StagedModelDataHandlerUtil.importStagedModel(
+			portletDataContext, exportedStagedModel);
 
 		validateCompanyDependenciesImport(dependentStagedModelsMap, liveGroup);
-	}
-
-	@Test
-	public void testExportImportFileExtension() throws Exception {
-		String sourceFileName = RandomTestUtil.randomString() + ".pdf";
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				stagingGroup.getGroupId(), TestPropsValues.getUserId());
-
-		FileEntry fileEntry = DLAppLocalServiceUtil.addFileEntry(
-			TestPropsValues.getUserId(), stagingGroup.getGroupId(),
-			DLFolderConstants.DEFAULT_PARENT_FOLDER_ID, sourceFileName,
-			ContentTypes.APPLICATION_PDF,
-			RandomTestUtil.randomBytes(TikaSafeRandomizerBumper.INSTANCE),
-			serviceContext);
-
-		exportImportStagedModel(fileEntry);
-
-		FileEntry importedFileEntry =
-			DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
-				fileEntry.getUuid(), liveGroup.getGroupId());
-
-		Assert.assertEquals(importedFileEntry.getExtension(), "pdf");
-
-		String title = RandomTestUtil.randomString() + ".awesome";
-
-		DLAppServiceUtil.updateFileEntry(
-			fileEntry.getFileEntryId(), StringPool.BLANK,
-			ContentTypes.TEXT_PLAIN, title, StringPool.BLANK, StringPool.BLANK,
-			false, (byte[])null, serviceContext);
-
-		exportImportStagedModel(fileEntry);
-
-		importedFileEntry = DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
-			fileEntry.getUuid(), liveGroup.getGroupId());
-
-		Assert.assertEquals(importedFileEntry.getExtension(), "pdf");
 	}
 
 	protected Map<String, List<StagedModel>> addCompanyDependencies()
@@ -169,19 +139,20 @@ public class FileEntryStagedModelDataHandlerTest
 		throws Exception {
 
 		Map<String, List<StagedModel>> dependentStagedModelsMap =
-			new LinkedHashMap<>();
+			new HashMap<>();
 
 		DDMStructure ddmStructure = DDMStructureTestUtil.addStructure(
 			group.getGroupId(), DLFileEntryMetadata.class.getName());
+
+		addDependentStagedModel(
+			dependentStagedModelsMap,
+			DDMStructureManagerUtil.getDDMStructureModelClass(), ddmStructure);
 
 		DLFileEntryType dlFileEntryType = addDLFileEntryType(
 			group.getGroupId(), ddmStructure.getStructureId());
 
 		addDependentStagedModel(
 			dependentStagedModelsMap, DLFileEntryType.class, dlFileEntryType);
-
-		addDependentStagedModel(
-			dependentStagedModelsMap, DDMStructure.class, ddmStructure);
 
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext(
@@ -233,33 +204,61 @@ public class FileEntryStagedModelDataHandlerTest
 			ServiceContextTestUtil.getServiceContext(
 				group.getGroupId(), TestPropsValues.getUserId());
 
-		DLAppTestUtil.populateServiceContext(
-			serviceContext, dlFileEntryType.getFileEntryTypeId());
-
-		return DLAppLocalServiceUtil.addFileEntry(
-			TestPropsValues.getUserId(), group.getGroupId(),
-			folder.getFolderId(), RandomTestUtil.randomString() + ".txt",
-			ContentTypes.TEXT_PLAIN,
-			RandomTestUtil.randomBytes(TikaSafeRandomizerBumper.INSTANCE),
+		folder = DLAppServiceUtil.addFolder(
+			group.getGroupId(), folder.getFolderId(),
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
 			serviceContext);
+
+		DLFolder dlFolder = (DLFolder)folder.getModel();
+
+		dlFolder.setDefaultFileEntryTypeId(
+			dlFileEntryType.getFileEntryTypeId());
+		dlFolder.setRestrictionType(
+			DLFolderConstants.RESTRICTION_TYPE_FILE_ENTRY_TYPES_AND_WORKFLOW);
+
+		DLFolderLocalServiceUtil.updateDLFolder(dlFolder);
+
+		List<Long> dlFileEntryTypeIds = new ArrayList<>();
+
+		dlFileEntryTypeIds.add(dlFileEntryType.getFileEntryTypeId());
+
+		DLFileEntryTypeLocalServiceUtil.updateFolderFileEntryTypes(
+			dlFolder, dlFileEntryTypeIds, dlFileEntryType.getFileEntryTypeId(),
+			serviceContext);
+
+		return folder;
 	}
 
 	@Override
-	protected StagedModel addVersion(StagedModel stagedModel) throws Exception {
-		FileEntry fileEntry = (FileEntry)stagedModel;
+	protected void deleteStagedModel(
+			StagedModel stagedModel,
+			Map<String, List<StagedModel>> dependentStagedModelsMap,
+			Group group)
+		throws Exception {
 
-		return DLAppServiceUtil.updateFileEntry(
-			fileEntry.getFileEntryId(), StringPool.BLANK,
-			ContentTypes.TEXT_PLAIN, RandomTestUtil.randomString(),
-			StringPool.BLANK, StringPool.BLANK, false, (byte[])null,
-			ServiceContextThreadLocal.getServiceContext());
+		List<StagedModel> fileEntryTypestagedModels =
+			dependentStagedModelsMap.get(DLFileEntryType.class.getSimpleName());
+
+		@SuppressWarnings("rawtypes")
+		StagedModelDataHandler stagedModelDataHandler =
+			StagedModelDataHandlerRegistryUtil.getStagedModelDataHandler(
+				DLFileEntryType.class.getName());
+
+		for (StagedModel fileEntryTypeStagedModel : fileEntryTypestagedModels) {
+			stagedModelDataHandler.deleteStagedModel(fileEntryTypeStagedModel);
+		}
+
+		super.deleteStagedModel(stagedModel, dependentStagedModelsMap, group);
 	}
 
 	@Override
 	protected StagedModel getStagedModel(String uuid, Group group) {
 		try {
-			return DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
-				uuid, group.getGroupId());
+			DLFolder dlFolder =
+				DLFolderLocalServiceUtil.getDLFolderByUuidAndGroupId(
+					uuid, group.getGroupId());
+
+			return new LiferayFolder(dlFolder);
 		}
 		catch (Exception e) {
 			return null;
@@ -268,17 +267,7 @@ public class FileEntryStagedModelDataHandlerTest
 
 	@Override
 	protected Class<? extends StagedModel> getStagedModelClass() {
-		return DLFileEntry.class;
-	}
-
-	@Override
-	protected boolean isCommentableStagedModel() {
-		return true;
-	}
-
-	@Override
-	protected boolean isVersionableStagedModel() {
-		return true;
+		return DLFolder.class;
 	}
 
 	protected void validateCompanyDependenciesImport(
@@ -311,7 +300,7 @@ public class FileEntryStagedModelDataHandlerTest
 			(DLFileEntryType)dlFileEntryTypesDependentStagedModels.get(0);
 
 		Assert.assertNull(
-			"Company DL file entry dependency should not be imported",
+			"Company DL file entry should not be imported",
 			DLFileEntryTypeLocalServiceUtil.
 				fetchDLFileEntryTypeByUuidAndGroupId(
 					dlFileEntryType.getUuid(), group.getGroupId()));
@@ -348,15 +337,15 @@ public class FileEntryStagedModelDataHandlerTest
 		DLFileEntryTypeLocalServiceUtil.getDLFileEntryTypeByUuidAndGroupId(
 			dlFileEntryType.getUuid(), group.getGroupId());
 
-		List<StagedModel> foldersDependentStagedModels =
+		List<StagedModel> folderDependentStagedModels =
 			dependentStagedModelsMap.get(DLFolder.class.getSimpleName());
 
-		Assert.assertEquals(1, foldersDependentStagedModels.size());
+		Assert.assertEquals(1, folderDependentStagedModels.size());
 
-		Folder folder = (Folder)foldersDependentStagedModels.get(0);
+		Folder parentFolder = (Folder)folderDependentStagedModels.get(0);
 
 		DLFolderLocalServiceUtil.getDLFolderByUuidAndGroupId(
-			folder.getUuid(), group.getGroupId());
+			parentFolder.getUuid(), group.getGroupId());
 	}
 
 	@Override
@@ -365,54 +354,20 @@ public class FileEntryStagedModelDataHandlerTest
 		throws Exception {
 
 		Assert.assertTrue(
-			String.valueOf(stagedModel.getCreateDate()) + StringPool.SPACE +
+			stagedModel.getCreateDate() + " " +
 				importedStagedModel.getCreateDate(),
 			DateUtil.equals(
 				stagedModel.getCreateDate(),
 				importedStagedModel.getCreateDate(), true));
-
 		Assert.assertEquals(
 			stagedModel.getUuid(), importedStagedModel.getUuid());
 
-		FileEntry fileEntry = (FileEntry)stagedModel;
-		FileEntry importedFileEntry = (FileEntry)importedStagedModel;
+		Folder folder = (Folder)stagedModel;
+		Folder importedFolder = (Folder)importedStagedModel;
 
+		Assert.assertEquals(folder.getName(), importedFolder.getName());
 		Assert.assertEquals(
-			fileEntry.getFileName(), importedFileEntry.getFileName());
-		Assert.assertEquals(
-			fileEntry.getExtension(), importedFileEntry.getExtension());
-		Assert.assertEquals(
-			fileEntry.getMimeType(), importedFileEntry.getMimeType());
-		Assert.assertEquals(fileEntry.getTitle(), importedFileEntry.getTitle());
-		Assert.assertEquals(
-			fileEntry.getDescription(), importedFileEntry.getDescription());
-		Assert.assertEquals(fileEntry.getSize(), importedFileEntry.getSize());
-
-		FileVersion latestFileVersion = fileEntry.getLatestFileVersion();
-		FileVersion importedLatestFileVersion =
-			importedFileEntry.getLatestFileVersion();
-
-		Assert.assertEquals(
-			latestFileVersion.getUuid(), importedLatestFileVersion.getUuid());
-		Assert.assertEquals(
-			latestFileVersion.getFileName(),
-			importedLatestFileVersion.getFileName());
-		Assert.assertEquals(
-			latestFileVersion.getExtension(),
-			importedLatestFileVersion.getExtension());
-		Assert.assertEquals(
-			latestFileVersion.getMimeType(),
-			importedLatestFileVersion.getMimeType());
-		Assert.assertEquals(
-			latestFileVersion.getTitle(), importedLatestFileVersion.getTitle());
-		Assert.assertEquals(
-			latestFileVersion.getDescription(),
-			importedLatestFileVersion.getDescription());
-		Assert.assertEquals(
-			latestFileVersion.getSize(), importedLatestFileVersion.getSize());
-		Assert.assertEquals(
-			latestFileVersion.getStatus(),
-			importedLatestFileVersion.getStatus());
+			folder.getDescription(), importedFolder.getDescription());
 	}
 
 }
